Cleanup Examples, add Readme to run examples
