+++ .gitignore
@@ -5,7 +5,6 @@ cp_plugin.sh
 zenoh-ts/dist
 zenoh-ts/node_modules
 zenoh-ts/package-lock.json
-zenoh-ts/examples/node/src/index.js
 zenoh-ts/examples/web/src/index.js
 zenoh-ts/cjs/
 zenoh-ts/esm/
@@ -13,3 +12,4 @@ zenoh-ts/esm/
 zenoh-plugin-remote-api/target/** 
 zenoh-plugin-remote-api/bindings/** 
 **/target/** 
+**/node_modules/**
