+++ zenoh-plugin-remote-api/EXAMPLE_CONFIG.json5
@@ -2,14 +2,11 @@
   mode: "router",
   plugins_loading: {
      enabled: true,
-     search_dirs: ["../target/debug", "~/.zenoh/lib"],
+     search_dirs: ["./target/debug", "~/.zenoh/lib"],
    },
   plugins: {
     remote_api: {
         "websocket_port": "10000",
-    },
-    rest:{
-	"http_port":"8080"
     },
   },
 }
