+++ zenoh-ts/examples/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite --config vite.config.js",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "clean": "rm -rf ./node_modules ./dist ./esm"
   },
   "devDependencies": {
     "@types/uuid": "^10.0.0",
@@ -14,9 +15,8 @@
     "vite": "^5.2.12"
   },
   "dependencies": {
-    "@ZettaScaleLabs/zenoh-ts": "file:../../",
+    "@ZettaScaleLabs/zenoh-ts": "file:../",
     "eslint": "^9.10.0",
-    "jscdr": "0.0.3",
     "tslog": "^4.9.3"
   }
 }
