--- zenoh-ts/examples/ros2_teleop/index.html
@@ -1,194 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
-    />
-    <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  </head>
-
-  <body class="w3-container">
-    <script type="module" src="/src/main.ts"></script>
-    <!-- <script src="https://cdn.jsdelivr.net/npm/bytebuffer@5.0.1"></script>
-  <script src="https://cdn.jsdelivr.net/npm/jscdr@0.0.1"></script> -->
-
-    <div class="w3-container w3-bar w3-card-4 w3-green w3-margin-bottom">
-      <h3 class="w3-bar-item">Zenoh ROS2 teleop</h3>
-      <h3 id="battery_label" class="w3-bar-item w3-right">Battery: - %</h3>
-    </div>
-
-    <div class="w3-row">
-      <div class="w3-col s2 m2 l2">
-        <label for="locator">Zenoh locator:</label>
-      </div>
-      <div class="w3-col s5 m5 l5">
-        <input
-          id="locator"
-          class="w3-input w3-border w3-padding-small"
-          type="text"
-          required
-          value="ws/127.0.0.1:10000"
-        />
-      </div>
-    </div>
-    <div class="w3-row">
-      <div class="w3-col s2 m2 l2">
-        <label for="locator">Ros2 Scope : </label>
-      </div>
-      <div class="w3-col s5 m5 l5">
-        <input
-          id="scope_input"
-          class="w3-input w3-border w3-padding-small w3-margin-right"
-          type="text"
-          required
-          value="bot"
-        />
-      </div>
-    </div>
-
-    <br />
-
-    <!-- If your robot has a camera and zcapture installed (from zenoh-demos/computer-vision/zcam/), uncomment this: 
-    
-    <div class="w3-card-4 w3-margin-bottom">
-        <header class="w3-bar w3-green" onclick="document.getElementById('Camera').classList.toggle('w3-hide');">
-            <h5 id="camera_label" class="w3-bar-item" style="margin: 0;">Camera</h5>
-            <h5 class="w3-bar-item w3-right" style="margin: 0;"><i class='fa fa-video-camera'></i></h5>
-        </header>
-        <div id="Camera" class="w3-container w3-padding">
-            <img id="camera_img" class="w3-container w3-padding" style="max-width: 95%; max-height: 300px;" src="" />
-        </div>
-        <br />
-    </div>
-    -->
-
-    <div class="w3-card-4 w3-margin-bottom">
-      <header
-        class="w3-bar w3-green"
-        onclick="document.getElementById('Drive').classList.toggle('w3-hide');"
-      >
-        <h5 id="drive_label" class="w3-bar-item" style="margin: 0">Drive</h5>
-        <h5 class="w3-bar-item w3-right" style="margin: 0">
-          <i class="fa fa-gamepad"></i>
-        </h5>
-      </header>
-      <div id="Drive" class="w3-container w3-padding">
-        <div class="w3-auto" style="display: grid; width: 12em; height: 9em">
-          <button
-            onmousedown="pubTwist(1.0, 0.0);"
-            onmouseup="pubTwist(0.0, 0.0);"
-            ontouchstart="pubTwist(1.0, 0.0);"
-            ontouchend="pubTwist(0.0, 0.0);"
-            style="font-size: 2em; grid-column: 2; grid-row: 1"
-          >
-            <i class="fas fa-caret-up"></i>
-          </button>
-          <button
-            onmousedown="pubTwist(0.0, 1.0);"
-            onmouseup="pubTwist(0.0, 0.0);"
-            ontouchstart="pubTwist(0.0, 1.0);"
-            ontouchend="pubTwist(0.0, 0.0);"
-            style="font-size: 2em; grid-column: 1; grid-row: 2"
-          >
-            <i class="fas fa-caret-left"></i>
-          </button>
-          <button
-            onmousedown="pubTwist(-1.0, 0.0);"
-            onmouseup="pubTwist(0.0, 0.0);"
-            ontouchstart="pubTwist(-1.0, 0.0);"
-            ontouchend="pubTwist(0.0, 0.0);"
-            style="font-size: 2em; grid-column: 2; grid-row: 2"
-          >
-            <i class="fas fa-caret-down"></i>
-          </button>
-          <button
-            onmousedown="pubTwist(0.0, -1.0);"
-            onmouseup="pubTwist(0.0, 0.0);"
-            ontouchstart="pubTwist(0.0, -1.0);"
-            ontouchend="pubTwist(0.0, 0.0);"
-            style="font-size: 2em; grid-column: 3; grid-row: 2"
-          >
-            <i class="fas fa-caret-right"></i>
-          </button>
-          <button
-            onclick="pubTwist(0.0, 0.0);"
-            style="font-size: 2em; grid-column: 1 / 4; grid-row: 3"
-          >
-            STOP
-          </button>
-        </div>
-        <br />
-      </div>
-    </div>
-
-    <div class="w3-card-4 w3-margin-bottom">
-      <header
-        class="w3-bar w3-green"
-        onclick="document.getElementById('Lidar').classList.toggle('w3-hide');"
-      >
-        <h5 id="lidar_label" class="w3-bar-item" style="margin: 0">Lidar</h5>
-        <h5 class="w3-bar-item w3-right" style="margin: 0">
-          <i class="fa fa-line-chart"></i>
-        </h5>
-      </header>
-      <div
-        id="Lidar"
-        class="w3-container w3-padding"
-        style="width: 95%; height: 300px"
-      >
-        <canvas id="Lidar-canvas" width="300" height="300"></canvas>
-      </div>
-      <br />
-    </div>
-
-    <div class="w3-card-4 w3-margin-bottom">
-      <header
-        class="w3-bar w3-green"
-        onclick="document.getElementById('Logs').classList.toggle('w3-hide');"
-      >
-        <h5 id="logs_label" class="w3-bar-item" style="margin: 0">Logs</h5>
-        <h5 class="w3-bar-item w3-right" style="margin: 0">
-          <i class="fa fa-envelope"></i>
-        </h5>
-      </header>
-      <div id="Logs" class="w3-container w3-padding">
-        <div
-          style="overflow: auto; height: 200px; border: 1px solid black"
-          id="rosout_logs"
-        ></div>
-      </div>
-    </div>
-
-    <div class="w3-card-4 w3-margin-bottom">
-      <header
-        class="w3-bar w3-green"
-        onclick="document.getElementById('Config').classList.toggle('w3-hide');"
-      >
-        <h5 class="w3-bar-item" style="margin: 0">Config</h5>
-        <h5 class="w3-bar-item w3-right" style="margin: 0">
-          <i class="fa fa-gear"></i>
-        </h5>
-      </header>
-      <form id="Config" class="w3-container w3-hide">
-        <label for="linear_scale">Linear scale:</label>
-        <input
-          id="linear_scale"
-          class="w3-input w3-border w3-padding-small"
-          type="number"
-          required
-          value="10"
-        /><br />
-        <label for="angular_scale">Angular scale:</label>
-        <input
-          id="angular_scale"
-          class="w3-input w3-border w3-padding-small"
-          type="number"
-          required
-          value="20"
-        /><br />
-      </form>
-    </div>
-  </body>
-</html>
