--- zenoh-ts/examples/ros2_teleop/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "zenoh-ts",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite --config vite.config.js",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "@types/uuid": "^10.0.0",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.12"
-  },
-  "dependencies": {
-    "@types/bytebuffer": "^5.0.49",
-    "bytebuffer": "^5.0.1",
-    "jscdr": "^0.0.3",
-    "tslog": "^4.9.3",
-    "@ZettaScaleLabs/zenoh-ts": "file:../../"
-  }
-}
