--- zenoh-ts/examples/ros2_teleop/vite.config.js
@@ -1,16 +1,0 @@
-import { defineConfig } from "vite";
-
-// vite.config.js
-export default defineConfig({
-  build: {
-    outDir: "./dist",
-    rollupOptions: {
-      input: {
-        app: "./ros2-teleop.html",
-      },
-    },
-    server: {
-      open: "./ros2-teleop.html",
-    },
-  },
-});
