+++ zenoh-ts/examples/src/vite-env.d.ts
