+++ zenoh-ts/examples/src/z_delete.ts
@@ -1,8 +1,10 @@
 import "./style.css";
 import "./webpage.ts";
+
 import { Config, Session } from "@ZettaScaleLabs/zenoh-ts";
 
-export async function main_put() {
+export async function _delete() {
   const session = await Session.open(Config.new("ws/127.0.0.1:10000"));
-  session.put("demo/example/zenoh-ts-put", "Put from Typescript!");
+  session.delete("demo/example/zenoh-ts-delete");
+  await session.close();
 }
