+++ zenoh-ts/examples/src/z_put.ts
@@ -1,10 +1,8 @@
 import "./style.css";
 import "./webpage.ts";
-
 import { Config, Session } from "@ZettaScaleLabs/zenoh-ts";
 
-export async function main_delete() {
+export async function put() {
   const session = await Session.open(Config.new("ws/127.0.0.1:10000"));
-  session.delete("demo/example/zenoh-ts-delete");
-  await session.close();
+  session.put("demo/example/zenoh-ts-put", "Put from Typescript!");
 }
