+++ zenoh-ts/examples/tsconfig.json
@@ -8,15 +8,14 @@
 
     /* Bundler mode */
     "moduleResolution": "bundler",
-    "allowJs": true,
     "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "esModuleInterop": true,
+
     /* Linting */
     "strict": true,
-    "noUnusedLocals": true,
+    "noUnusedLocals": false,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
