+++ zenoh-ts/package.json
@@ -17,11 +17,7 @@
   ],
   "dependencies": {
     "channel-ts": "^0.1.2",
-    "cpx": "^1.5.0",
     "eslint": "^9.10.0",
-    "fp-ts": "^2.16.5",
-    "object-inspect": "^1.13.1",
-    "request": "^2.79.0",
     "text-encoding": "^0.7.0",
     "ts-log": "^2.2.5",
     "tslog": "^4.9.3",
@@ -29,13 +25,9 @@
   },
   "devDependencies": {
     "@types/node": "20.9.1",
-    "@types/object-inspect": "^1.13.0",
     "@types/uuid": "^10.0.0",
-    "ts-loader": "^9.5.0",
-    "ts-node": "10.9.1",
     "typedoc": "^0.26.5",
-    "typescript": "^5.3.2",
-    "webpack-cli": "^5.1.4"
+    "typescript": "^5.3.2"
   },
   "scripts": {
     "build": "npx tsc",
